% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\name{pmc_to_markdown}
\alias{pmc_to_markdown}
\title{Fetch a PMC article rendered as Markdown}
\usage{
pmc_to_markdown(client, pmcid)
}
\arguments{
\item{client}{A \code{pubmed_client} created by \code{\link[=pubmed_client]{pubmed_client()}}.}

\item{pmcid}{A PMC identifier, e.g. \code{"PMC7906746"}.}
}
\value{
A length-one character vector containing the Markdown document.
}
\description{
Fetch a PMC article rendered as Markdown
}
