% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\name{pubmed_fetch}
\alias{pubmed_fetch}
\title{Fetch article metadata}
\usage{
pubmed_fetch(client, pmids)
}
\arguments{
\item{client}{A \code{pubmed_client} created by \code{\link[=pubmed_client]{pubmed_client()}}.}

\item{pmids}{One or more PMIDs.}
}
\value{
For a single PMID, a named list of article fields. For several PMIDs,
a list of such named lists.
}
\description{
Fetch article metadata
}
