% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\name{pubmed_search}
\alias{pubmed_search}
\title{Search PubMed}
\usage{
pubmed_search(client, query, limit = 20L)
}
\arguments{
\item{client}{A \code{pubmed_client} created by \code{\link[=pubmed_client]{pubmed_client()}}.}

\item{query}{PubMed search query string.}

\item{limit}{Maximum number of PMIDs to return.}
}
\value{
A character vector of PMIDs.
}
\description{
Search PubMed
}
