% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\name{pubmed_search_and_fetch}
\alias{pubmed_search_and_fetch}
\title{Search PubMed and fetch metadata in one call}
\usage{
pubmed_search_and_fetch(client, query, limit = 20L)
}
\arguments{
\item{client}{A \code{pubmed_client} created by \code{\link[=pubmed_client]{pubmed_client()}}.}

\item{query}{PubMed search query string.}

\item{limit}{Maximum number of articles to fetch.}
}
\value{
A list of named lists, one per article.
}
\description{
Search PubMed and fetch metadata in one call
}
