% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\docType{package}
\name{pubmedclient-package}
\alias{pubmedclient}
\alias{pubmedclient-package}
\title{pubmedclient: PubMed and PMC API Client}
\description{
R bindings for the Rust \code{pubmed-client} library. Provides access to
the PubMed and PubMed Central (PMC) APIs for searching, fetching article
metadata, and retrieving full text or Markdown, and to the Europe PMC REST
API for cross-source search and citation graphs. Wraps the high-performance
asynchronous Rust client behind a synchronous R interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/illumination-k/pubmed-client}
  \item Report bugs at \url{https://github.com/illumination-k/pubmed-client/issues}
}

}
\author{
\strong{Maintainer}: illumination-k \email{illumination27@gmail.com}
}
\keyword{internal}
