% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed-client.R
\docType{package}
\name{pubmedclient-package}
\alias{pubmedclient}
\alias{pubmedclient-package}
\title{pubmedclient: PubMed and PMC API Client}
\description{
R bindings for the Rust \code{pubmed-client} library. Search PubMed, fetch article
metadata, and retrieve PMC full text or Markdown.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/illumination-k/pubmed-client}
  \item Report bugs at \url{https://github.com/illumination-k/pubmed-client/issues}
}

}
\author{
\strong{Maintainer}: illumination-k \email{illumination27@gmail.com}
}
\keyword{internal}
